% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bayesfertility}
\alias{bayesfertility}
\title{bayesfertility: Fit Bayesian fertility schedule models}
\description{
Fit Bayesian fertility schedule models, including the Peristera-Kostaki model
}
\concept{help}
