% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_formula.R
\name{extract_parametric_terms}
\alias{extract_parametric_terms}
\title{Build a list of parametric terms from a model matrix and formula}
\usage{
extract_parametric_terms(X, terms_obj)
}
\arguments{
\item{X}{model matrix (with "assign" attribute) or design matrix from jagam}

\item{formula}{or pterms object the columns came from}
}
\value{
named list: term label -> integer vector of column indices in X
}
\description{
Build a list of parametric terms from a model matrix and formula
}
