% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_fit.R
\name{pk_fit}
\alias{pk_fit}
\title{Fit the Peristera-Kostaki Model}
\usage{
pk_fit(
  data,
  births,
  exposure,
  age,
  effects = pk_effects(),
  sample_prior = c("no", "yes", "only"),
  ...
)
}
\arguments{
\item{data}{data frame}

\item{births}{column name of number of births}

\item{exposure}{column name of exposure}

\item{age}{column name of age}

\item{effects}{model specification, specify via \code{pk_effects()}}
}
\description{
Fit the Peristera-Kostaki Model
}
