% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_tfr.R
\name{posterior_tfr}
\alias{posterior_tfr}
\title{Calculate posterior TFR}
\usage{
posterior_tfr(
  object,
  newdata = NULL,
  age_range = c(15, 49),
  age_step = 0.5,
  source = c("prior", "posterior"),
  summarize = TRUE,
  conf.level = 0.95,
  ndraws = NULL,
  ...
)
}
\arguments{
\item{object}{object}
}
\description{
Calculate posterior TFR
}
