% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_posterior_predict.R, R/pk_pp_check.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{posterior_predict}
\alias{pp_check}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayesplot}{\code{\link[bayesplot]{pp_check}}}

  \item{rstantools}{\code{\link[rstantools]{posterior_predict}}}
}}

